+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 * [IN PROGRESS] C MEX API:
 
+  * mxGetLogicals, mxIsLogicalScalar, mxIsLogicalScalarTrue functions.
+
+  * mxGetInt8s, mxSetInt8s, mxGetComplexInt8s, mxSetComplexInt8s, mxGetUint8s, mxSetUint8s, mxGetComplexUint8s
+    mxSetComplexUint8s, mxGetInt16s, mxSetInt16s, mxGetComplexInt16s, mxSetComplexInt16s, mxGetUint16s, mxGetComplexUint16s
+    mxSetComplexUint16s, mxGetInt32s, mxSetInt32s, mxGetComplexInt32s, mxSetComplexInt32s, mxGetUint32s, mxSetUint32s
+    mxGetComplexUint32s, mxSetComplexUint32s, mxSetUint16s, mxGetInt64s, mxSetInt64s, mxGetComplexInt64s, mxSetComplexInt64s
+    mxGetUint64s, mxSetUint64s, mxGetComplexUint64s, mxSetComplexUint64s functions.
+
   * mxIsObject, mxIsFunctionHandle, mxIsOpaque functions.
 
   * mxIsInt8, mxIsInt16, mxIsInt32, mxIsInt64, mxIsUint8, mxIsUint16, mxIsUint32, mxIsUint64 functions.